+++ inc/lilyParser/index.ts
@@ -1,0 +1,16 @@
+
+
+//import lilyGrammar from "./lilypond.jison";
+//console.log("lilyGrammar:", lilyGrammar);
+
+const jison = require("jison");
+
+
+
+const createParser = grammar => new jison.Parser(grammar);
+
+
+
+export {
+	createParser,
+};