+++ changelog.md
@@ -12,6 +12,11 @@ TODO:
 
 
 
+*	2020.4.14
+
+	+	Lilypond parser based on jison created.
+
+
 *	2020.4.13
 
 	+	lilyParser.ts: escapeComments added.